Corrected PRORWTS2 source c/o @qkumba
This has the support for scenery disks, and incorporates the banking
fix from 2a98378. The disk image in orig/ is updated with this new
build of FS2. The binary builds to a slightly different start location
and one intermediate file is no longer produced, necessitating some
script and Makefile updates.